making Request function separated
