+++ src/main/java/Functions/FileBytes.java
@@ -24,16 +24,12 @@ public class FileBytes {
 
                     content.append(line).append(System.lineSeparator());
                     System.out.println("line:  " + line);
-                    //line.contains("<a href=\"index.html\" class=\"navbar-brand\">SLiPP</a>")
                     if (user != null && line.contains("<ul class=\"nav navbar-nav navbar-right\">") && !imbed) {
-                        System.out.println("User name : " + user.getName());
-                        //content.append("<li class=\"name\">");
                         content.append("<li>");
                         content.append("<div class=\"name\">" + user.getName() + " 님 " +  "</div>");
                         content.append("</li>");
                         imbed = true;
                     }
-
 
                 }
                 return content.toString().getBytes("UTF-8");
@@ -44,18 +40,14 @@ public class FileBytes {
 
         else{
 
-
             try (FileInputStream fis = new FileInputStream(fileURL);
                  ByteArrayOutputStream bos = new ByteArrayOutputStream()) {
 
                 byte[] buffer = new byte[1024];
                 int bytesRead;
                 while ((bytesRead = fis.read(buffer)) != -1) {
-
                     bos.write(buffer, 0, bytesRead);
-                    //System.out.println("line:" + new String(buffer, 0, bytesRead, "UTF-8"));
                 }
-
                 return body = bos.toByteArray();
             } catch (IOException e) {
                 e.printStackTrace();
@@ -63,27 +55,6 @@ public class FileBytes {
 
 
         }
-
-        /*
-        byte[] body = new byte[0];
-
-        try (FileInputStream fis = new FileInputStream(fileURL);
-             ByteArrayOutputStream bos = new ByteArrayOutputStream()) {
-
-            byte[] buffer = new byte[1024];
-            int bytesRead;
-            while ((bytesRead = fis.read(buffer)) != -1) {
-
-                bos.write(buffer, 0, bytesRead);
-                //System.out.println("line:" + new String(buffer, 0, bytesRead, "UTF-8"));
-            }
-
-            return body = bos.toByteArray();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-         */
 
         return body;
     }
