+++ src/main/java/HTTPModel/HttpStatus.java
@@ -1,4 +1,4 @@
-package DTO;
+package HTTPModel;
 
 public enum HttpStatus {
     OK(200, "OK"),
