+++ src/main/java/HTTPModel/Response.java
@@ -1,4 +1,4 @@
-package DTO;
+package HTTPModel;
 
 import java.util.Random;
